refactor: synched functions and web-client models
