+++ functions/src/models/Post/IPost.ts
@@ -5,7 +5,9 @@ export interface IPost extends firebase.firestore.DocumentData {
   postRef: string;
   isResponse: boolean;
   isRequest: boolean;
+  creatorRef: string;
   creatorSnapshot: IUser;
+  genericStatus: GenericPostStatus;
   title: string;
   description: string;
   streetAddress: string;
