--- functions/src/models/Post/UnauthenticatedPost.ts
@@ -1,245 +1,0 @@
-import {
-  Allow,
-  IsNotEmpty,
-  IsObject,
-  IsString,
-  ValidateNested,
-} from 'class-validator';
-import { firestore } from 'firebase';
-import { firebaseFirestore as db } from 'src/firebaseConfig';
-
-import { IStrippedUser } from '../users/IStrippedUser';
-import { ILatLngObject } from './ILatLngObject';
-import { IUnauthenticatedPost } from './IUnauthenticatedPost';
-import { Post } from './Post';
-
-export class UnauthenticatedPost implements IUnauthenticatedPost {
-  constructor(
-    postRef: string,
-    isRequest: boolean,
-    creatorSnapshot: IStrippedUser,
-    title: string,
-    description: string,
-    latLng: ILatLngObject,
-    createdAt = new Date(),
-    updatedAt = new Date(),
-  ) {
-    this._postRef = postRef;
-    this._isRequest = isRequest;
-    this._creatorSnapshot = creatorSnapshot;
-    this._title = title;
-    this._description = description;
-    this._latLng = latLng;
-    this._createdAt = createdAt;
-    this._updatedAt = updatedAt;
-  }
-
-  @ValidateNested()
-  private _postRef: string;
-
-  get postRef(): string {
-    return this._postRef;
-  }
-
-  set postRef(value: string) {
-    this._postRef = value;
-  }
-
-  @Allow()
-  private _isRequest: boolean;
-
-  get isRequest(): boolean {
-    return this._isRequest;
-  }
-
-  set isRequest(isRequest: boolean) {
-    this._isRequest = isRequest;
-  }
-
-  @IsObject()
-  private _creatorSnapshot: IStrippedUser;
-
-  get creatorSnapshot(): IStrippedUser {
-    return this._creatorSnapshot;
-  }
-
-  set creatorSnapshot(value: IStrippedUser) {
-    this._creatorSnapshot = value;
-  }
-
-  @IsString()
-  @IsNotEmpty()
-  private _title: string;
-
-  get title(): string {
-    return this._title;
-  }
-
-  set title(value: string) {
-    this._title = value;
-  }
-
-  @IsString()
-  private _description: string;
-
-  get description(): string {
-    return this._description;
-  }
-
-  set description(value: string) {
-    this._description = value;
-  }
-
-  @IsObject()
-  private _latLng: ILatLngObject;
-
-  get latLng(): ILatLngObject {
-    return this._latLng;
-  }
-
-  set latLng(value: ILatLngObject) {
-    this._latLng = value;
-  }
-
-  /* TODO: When we reach greater than 500 requests created per second:
-     https://firebase.google.com/docs/firestore/solutions/shard-timestamp#sharding_a_timestamp_field
-   */
-  @IsObject()
-  private _createdAt: Date;
-
-  get createdAt(): Date {
-    return this._createdAt;
-  }
-
-  set createdAt(value: Date) {
-    this._createdAt = value;
-  }
-
-  /* TODO: When we reach greater than 500 requests updated per second:
-     https://firebase.google.com/docs/firestore/solutions/shard-timestamp#sharding_a_timestamp_field
-   */
-  @IsObject()
-  private _updatedAt: Date;
-
-  get updatedAt(): Date {
-    return this._updatedAt;
-  }
-
-  set updatedAt(value: Date) {
-    this._updatedAt = value;
-  }
-
-  public static async fromPost(
-    data: Post,
-    path: string,
-  ): Promise<UnauthenticatedPost> {
-    return Promise.resolve(
-      new UnauthenticatedPost(
-        path,
-        data.isRequest,
-        {
-          displayName: data.creatorSnapshot.displayName || '',
-          displayPicture: data.creatorSnapshot.displayPicture,
-        },
-        data.title,
-        data.description,
-        {
-          latitude: data.latLng.latitude,
-          longitude: data.latLng.longitude,
-        },
-        data.createdAt.toDate(),
-        data.updatedAt.toDate(),
-      ),
-    );
-  }
-
-  public static fromFirestore(
-    data: firebase.firestore.DocumentData,
-  ): UnauthenticatedPost {
-    return new UnauthenticatedPost(
-      (data.postRef as firebase.firestore.DocumentReference).path,
-      data.isRequest,
-      data.userSnapshot,
-      data.title,
-      data.description,
-      {
-        latitude: (data.latLng as firebase.firestore.GeoPoint).latitude,
-        longitude: (data.latLng as firebase.firestore.GeoPoint).longitude,
-      },
-      data.createdAt.toDate(),
-      data.updatedAt.toDate(),
-    );
-  }
-
-  public static fromAlgolia(data: Record<string, any>): UnauthenticatedPost {
-    return new UnauthenticatedPost(
-      data.postRef,
-      data.isRequest,
-      data.pinUserSnapshot,
-      data.title,
-      data.description,
-      { latitude: data._geoloc.lat, longitude: data._geoloc.lng },
-      data.createdAt,
-      data.updatedAt,
-    );
-  }
-
-  public static fromObject(data: IUnauthenticatedPost): UnauthenticatedPost {
-    return new UnauthenticatedPost(
-      data.postRef,
-      data.isRequest,
-      data.creatorSnapshot,
-      data.title,
-      data.description,
-      data.latLng,
-      data.createdAt,
-      data.updatedAt,
-    );
-  }
-
-  toObject(): IUnauthenticatedPost {
-    return {
-      postRef: this.postRef,
-      isRequest: this.isRequest,
-      creatorSnapshot: this.creatorSnapshot,
-      title: this.title,
-      description: this.description,
-      latLng: this.latLng,
-      createdAt: this.createdAt,
-      updatedAt: this.updatedAt,
-    };
-  }
-
-  toFirestore(): firebase.firestore.DocumentData {
-    return {
-      postRef: db.doc(this.postRef),
-      isRequest: this.isRequest,
-      creatorSnapshot: this.creatorSnapshot,
-      title: this.title,
-      description: this.description,
-      latLng: new firestore.GeoPoint(
-        this.latLng.latitude,
-        this.latLng.longitude,
-      ),
-      createdAt: firestore.Timestamp.fromDate(this.createdAt),
-      updatedAt: firestore.Timestamp.fromDate(this.updatedAt),
-    };
-  }
-
-  toAlgolia(): object {
-    return {
-      postRef: this.postRef,
-      isRequest: this.isRequest,
-      objectID: db.doc(this.postRef).id,
-      creatorSnapshot: this.creatorSnapshot,
-      title: this.title,
-      description: this.description,
-      _geoloc: {
-        lat: this.latLng.latitude,
-        lng: this.latLng.longitude,
-      },
-      createdAt: this.createdAt,
-      updatedAt: this.updatedAt,
-    };
-  }
-}
