--- functions/src/models/users/IUserGeneral.ts
@@ -1,5 +1,0 @@
-export interface IUserGeneral {
-  displayName: string;
-  displayPicture: string | null;
-  rating: number | null;
-}
