+++ functions/src/models/users/User.ts
@@ -20,7 +20,7 @@ export class User implements IUser {
     casesCompleted = 0,
     postsMade = 0,
     displayNickname: string | null = null,
-    displayPicture?: string | null = null,
+    displayPicture: string | null = null,
     createdAt = firestore.Timestamp.now(),
   ) {
     this.postsMade = postsMade;
@@ -33,14 +33,7 @@ export class User implements IUser {
   }
 
   static factory = (data: IUser): User =>
-    new User(
-      data.username,
-      data.casesCompleted,
-      data.postsMade,
-      data.displayNickname,
-      data.displayPicture,
-      data.createdAt,
-    );
+    new User(data.username, data.casesCompleted, data.postsMade, data.displayNickname, data.displayPicture, data.createdAt);
 
   toObject = (): object => ({
     username: this.username,
